refactor(workflows): split lib/runtime into barrel + 6 seam modules
runtime.ts (1,634) becomes a pure re-export barrel over six cohesive
modules, preserving the public import path that index.ts, sibling lib
modules, and 6+ test files mock.module:

- instance-store.ts  — JSON-on-disk persistence + gate-notified tracking
- task-bridge.ts     — the single task-store/task-service crossing
- step-context.ts    — read-only info-gating query surface
- node-dispatch.ts   — system/plugin node execution + reopened-step dispatch
- engine.ts          — the mutual-recursion mutation knot (kept together)
- gates.ts           — human approve/reject/reopen

The three pure traversal helpers (flattenSteps/findStep/getTopLevelIndex)
move into the existing parser.ts. The only back-edge,
node-dispatch -> engine.completeStep, stays a dynamic import('./engine')
so the static graph is a DAG (gates -> engine -> node-dispatch -> leaves).
engine exports advanceWorkflow/propagateChildCompletion for gates, but the
barrel re-exports only today's 26 public names — public surface is
byte-identical.

Pure behavior-preserving relocation; the appendix's redesign opportunities
(decideGate dedup, atomic withInstance writes, contentDir-param drop,
require->import, discriminated getCurrentStep union) are deferred.

Verified: typecheck + eslint + workflows suite (419/0) + full suite
(5123/0) + binary build + isolated-home boot smoke (workflows activates,
GET /definitions and /instances return 200).
